+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.5.0] - Unreleased
+## [2.0.0] - Unreleased
+
 ### Added
 - Support of codeception v4
-
+- Add usage of codeception retry
 
 ## [1.4.0] - 2020-11-10